# swagger/swaggerDocs.yaml
swagger: "2.0"
info:
  description: "Monolith boilerplate API documentation."
  version: "1.0.0"
  title: "Monolith-Boilerplate"
basePath: "/"
...
schemes:
  - "http"
  - "https"
paths:
  /users:
    get:
      security:
        - ApiKeyAuth: []
        - RefreshAuth: []
      tags:
        - "user"
      summary: "Gets all users"
      description: "Retrieves all users from given database."
      operationId: "getAllUsers"
      produces:
        - "application/json"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/MoreUsersResponse"
        "404":
          description: "No user into the database"
          schema:
            type: "string"
            default: "No user into the database"
        "500":
          description: "Failure operation"
          schema:
            type: "string"
    post:
      security:
        - ApiKeyAuth: []
        - RefreshAuth: []
      tags:
        - "user"
      summary: "Create user"
      description: "Create a user and add it to the database"
      operationId: "createUser"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "user to add"
          required: true
          schema:
            $ref: "#/definitions/CreateUserRequest"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/UserSchema"
        "400":
          description: "Request has an invalid field"
          schema:
            type: "string"
            default: "Email already used"
        "500":
          description: "Failure operation"
          schema:
            type: "string"
            example: "Error during user creation"
  /users/{id}:
    get:
      security:
        - ApiKeyAuth: []
        - RefreshAuth: []
      tags:
        - "user"
      summary: "Get user by id"
      description: "Retrieves given user by its id from the database."
      operationId: "getUserByid"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "User Id."
          required: true
          type: "string"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/UserSchema"
        "400":
          description: "path parameter contains an invalid Id"
          schema:
            type: "string"
            default: "Invalid Id"
        "404":
          description: "User not found"
          schema:
            type: "string"
            default: "User not found"
        "500":
          description: "Failure operation"
          schema:
            type: "string"
    put:
      security:
        - ApiKeyAuth: []
        - RefreshAuth: []
      tags:
        - "user"
      summary: "Update user"
      description: "Update given user found by its id."
      operationId: "updateUser"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "User Id."
          required: true
          type: "string"
        - in: "body"
          name: "body"
          description: "Update for the user"
          required: true
          schema:
            $ref: "#/definitions/UserSchema"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/UserSchema"
        "400":
          description: "path parameter contains an invalid Id"
          schema:
            type: "string"
            default: "Invalid Id"
        "404":
          description: "User not found"
          schema:
            type: "string"
            default: "User not found"
        "500":
          description: "Failure operation"
          schema:
            type: "string"
    patch:
      security:
        - ApiKeyAuth: []
        - RefreshAuth: []
      tags:
        - "user"
      summary: "Patch user"
      description: "Edit some field of given user found by its id."
      operationId: "patchUser"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "User Id."
          required: true
          type: "string"
        - in: "body"
          name: "body"
          description: "Update for the user"
          required: true
          schema:
            $ref: "#/definitions/PatchUserRequest"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/UserSchema"
        "400":
          description: "path parameter contains an invalid Id"
          schema:
            type: "string"
            default: "Invalid Id"
        "404":
          description: "User not found"
          schema:
            type: "string"
            default: "User not found"
        "500":
          description: "Failure operation"
          schema:
            type: "string"
    delete:
      security:
        - ApiKeyAuth: []
        - RefreshAuth: []
      tags:
        - "user"
      summary: "Delete user"
      description: "Delete given user found by its id."
      operationId: "deleteUser"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "User Id."
          required: true
          type: "string"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/UserSchema"
        "400":
          description: "path parameter contains an invalid Id"
          schema:
            type: "string"
            default: "Invalid Id"
        "404":
          description: "User not found"
          schema:
            type: "string"
            default: "User not found"
        "500":
          description: "Failure operation"
          schema:
            type: "string"
  /users/blockUnblock/{id}:
    put:
      security:
        - ApiKeyAuth: []
        - RefreshAuth: []
      tags:
        - "user"
      summary: "Blocks or unblocks user"
      description: "Update given user found by its id, setting its status to 'blocked' or 'active' based on the action that needs to be done."
      operationId: "blockUnblockUser"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "User Id."
          required: true
          type: "string"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/UserSchema"
        "400":
          description: "path parameter contains an invalid Id"
          schema:
            type: "string"
            default: "Invalid Id"
        "404":
          description: "User not found"
          schema:
            type: "string"
            default: "User not found"
        "500":
          description: "Failure operation"
          schema:
            type: "string"
  /users/adminFilter:
    post:
      security:
        - ApiKeyAuth: []
        - RefreshAuth: []
      tags:
        - "user"
      summary: "Filter users for admin page"
      description: "Filters users into given database following some criteria"
      operationId: "adminUsersFilter"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "filter object"
          required: true
          schema:
            type: "object"
            properties:
              email:
                type: "string"
                example: "some-email@gmail.com"
              lastname:
                type: "string"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/MoreUsersResponse"
        "400":
          description: "Request has an invalid field"
          schema:
            type: "string"
            default: "The body field 'email' must be string"
        "500":
          description: "Failure operation"
          schema:
            type: "string"
  /users/changePassword:
    patch:
      security:
        - ApiKeyAuth: []
        - RefreshAuth: []
      tags:
        - "user"
      summary: "Change password"
      description: "Changes password of the user within a specific token. Used for password recover."
      operationId: "changePassword"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "new password"
          required: true
          schema:
            type: "object"
            properties:
              passowrd:
                type: "string"
                example: "newPassword1"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "string"
            example: "password changed"
        "401":
          description: "ResetPassword token invalid or expired"
          schema:
            type: "string"
            default: "ResetPassword token not valid"
        "404":
          description: "User not found"
          schema:
            type: "string"
            default: "User not found"
        "500":
          description: "Failure operation"
          schema:
            type: "string"
...
  /auth/user:
    get:
      security:
        - ApiKeyAuth: []
        - RefreshAuth: []
      tags:
        - "auth"
      summary: "Gets user into the token from request headers"
      description: "This api is used to retrieve a user from the database by its id present into token in request headers. It's needed to retrieve the authenticated user for frontend side, on page refresh."
      operationId: "getUserFromToken"
      produces:
        - "application/json"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/UserSchema"
        "404":
          description: "User not found"
          schema:
            type: "string"
            default: "User not found"
        "500":
          description: "Failure operation"
          schema:
            type: "string"
...
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: Authorization
  RefreshAuth:
    type: apiKey
    in: header
    name: refresh
security:
  - ApiKeyAuth: []
  - RefreshAuth: []
definitions:
  UserSchema:
    type: "object"
    properties:
      _id:
        type: "string"
      email:
        type: "string"
        format: "email"
      password:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      role:
        type: "string"
        enum: ["admin", "user"]
        default: "user"
      status:
        type: "string"
        enum: ["active", "expired"]
        default: "active"
      lastLogin:
        type: "string"
        format: "date-time"
  CreateUserRequest:
    type: "object"
    properties:
      email:
        type: "string"
        format: "email"
      password:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
...
  PatchUserRequest:
    type: "object"
    properties:
      email:
        type: "string"
        format: "email"
      firstName:
        type: "string"
      lastName:
        type: "string"
...
  MoreUsersResponse:
    type: "array"
    items:
      $ref: "#/definitions/UserSchema"
...